Update to newer version of deploy.cmd from kuduscript
Update to a newer version of `deploy.cmd` from kuduscript. Now it
no longer relies on a hardcoded version of npm (which was causing deployment
failures as the version was removed from web hosts).

Some script changes were necessary, namely that the script tries to copy `source`
to `target`, then run `npm install`, wherease we want to run `npm install` and
`generate` in `source`, then copy the output of the `out` folder to `target`.